与v307进行持续集成

+++ drivers/VirtualIO/TTP229/TTP229.cpp
@@ -7,12 +7,13 @@ void TTP229::update(){
     uint16_t new_map = 0;
     sck_gpio_.clr();
 
-    for (int i = 0; i < 16; i++)
-    {
+    for (size_t i = 0; i < 16; i++){
         sck_gpio_.set();
         sck_gpio_.clr();
-        new_map <<= 1; new_map |= sdo_gpio_.read().to_bool();
+        new_map <<= 1; 
+        new_map |= sdo_gpio_.read().to_bool();
     }
+
     map_ = new_map;
     uint8_t new_num = 0;
     while(new_map != (1 << new_num)){
@@ -25,4 +26,5 @@ void TTP229::update(){
     num_ = Some(new_num);
 }
 
-bool TTP229::is_idle(){return sdo_gpio_.read() == LOW;}
+bool TTP229::is_idle(){
+    return sdo_gpio_.read() == LOW;}
